migrate account balances to users controller
